djninja/requirements.txt
anyascii==0.3.0
asgiref==3.4.1
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
decorator==5.1.0
Django==3.2.9
django-appconf==1.0.5
django-compressor==2.4.1
django-debug-toolbar==3.2.2
django-extensions==3.1.5
django-filter==21.1
django-modelcluster==5.2
django-ninja==0.16.1
django-sass-processor==1.1
django-taggit==1.5.1
django-treebeard==4.5.1
djangorestframework==3.12.4
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==3.3
install==1.3.4
ipython==7.29.0
jedi==0.18.1
l18n==2021.3
libsass==0.21.0
matplotlib-inline==0.1.3
openpyxl==3.0.9
parso==0.8.2
phonenumbers==8.12.37
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.22
pydantic==1.8.2
Pygments==2.10.0
pytz==2021.3
rcssmin==1.0.6
requests==2.26.0
rjsmin==1.1.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tablib==3.1.0
telepath==0.2
traitlets==5.1.1
typing_extensions==4.0.0
urllib3==1.26.7
wagtail==2.15.1
wcwidth==0.2.5
webencodings==0.5.1
Willow==1.4
wincertstore==0.2
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
